requirements.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
sqlmodel<0.0.9
python-jose==3.2.0
cryptography<3.4
passlib==1.7.4
cachetools==4.2.4
python-multipart==0.0.5
psycopg2-binary==2.9.5
pydantic==1.8.2
